api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
